backend_django/requirements.txt
# Django Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery + Redis
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0
redis>=5.0,<6.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Claude AI
anthropic>=0.39,<1.0

# File handling
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
Pillow>=10.0,<11.0
python-magic>=0.4,<1.0

# Document processing
python-docx>=1.1,<2.0
mammoth>=1.8,<2.0
ebooklib>=0.18,<1.0

# ONIX / ISBN
lxml>=5.1,<6.0

# Payments
stripe>=7.0,<8.0

# Utilities
python-dateutil>=2.8,<3.0
uuid>=1.30
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0

# Dev
django-debug-toolbar>=4.3,<5.0
django-extensions>=3.2,<4.0
